REF: Refactor dMRI b0 handling of externally set multi-b0 data
Refactor dMRI b0 handling of externally set multi-b0 data. When the DWI
data class `bzero` attribute is set by the user to a 4D sequence, it is
set "as-is", as opposed to the case where multiple b0s are contained by
the primary DWI 4D sequence where the median is computed. Thus, in the
former case 4D data is held by the `bzero` attribute, whereas in the
latter, 3D data is held.

This makes such that when serializing a DWI instance, and depending on
whether the b0s are to be prepended to the DWI non-zero gradient data,
the concatenation fails for 4D b0 data.

WIP possibilities:
- Raise a validator error if 4D `bzero` data is provided.
- Warn the user and compute the median.

Explain the behavior in the documentation.